@@ -45,7 +45,7 @@ func TestARSAreCreated(t *testing.T) {
 		apiExportName = "example.com"
 	)
 
-	ctx := context.Background()
+	ctx := t.Context()
 	ctrlruntime.SetLogger(logr.Discard())
 
 	// setup a test environment in kcp
@@ -116,7 +116,7 @@ func TestARSAreNotUpdated(t *testing.T) {
 		apiExportName = "example.com"
 	)
 
-	ctx := context.Background()
+	ctx := t.Context()
 	ctrlruntime.SetLogger(logr.Discard())
 
 	// setup a test environment in kcp
@@ -166,14 +166,20 @@ func TestARSAreNotUpdated(t *testing.T) {
 		t.Fatalf("Failed to wait for APIResourceSchema to be created: %v", err)
 	}
 
-	if err := envtestClient.Get(ctx, ctrlruntimeclient.ObjectKeyFromObject(pr), pr); err != nil {
-		t.Fatalf("Failed to fetch PublishedResource: %v", err)
+	// prevent race condition of the controller first creating the ARS, then updating the PR,
+	// by waiting again for the status to be updated.
+	err = wait.PollUntilContextTimeout(ctx, 500*time.Millisecond, 30*time.Second, false, func(ctx context.Context) (done bool, err error) {
+		if err := envtestClient.Get(ctx, ctrlruntimeclient.ObjectKeyFromObject(pr), pr); err != nil {
+			return false, nil
+		}
+
+		return pr.Status.ResourceSchemaName != "", nil
+	})
+	if err != nil {
+		t.Fatalf("Failed to wait for PublishedResource status to be updated: %v", err)
 	}
 
 	arsName := pr.Status.ResourceSchemaName
-	if arsName == "" {
-		t.Fatal("Expected PublishedResource status to contain ARS name, but value is empty.")
-	}
 
 	// update the CRD
 	t.Logf("Updating CRD (same version, but new schema)…")
@@ -194,17 +200,20 @@ func TestARSAreNotUpdated(t *testing.T) {
 		t.Fatalf("Failed to wait for 2nd APIResourceSchema to be created: %v", err)
 	}
 
-	if err := envtestClient.Get(ctx, ctrlruntimeclient.ObjectKeyFromObject(pr), pr); err != nil {
-		t.Fatalf("Failed to fetch PublishedResource: %v", err)
+	// wait for the status to contain the new ARS name
+	err = wait.PollUntilContextTimeout(ctx, 500*time.Millisecond, 1*time.Minute, false, func(ctx context.Context) (done bool, err error) {
+		if err := envtestClient.Get(ctx, ctrlruntimeclient.ObjectKeyFromObject(pr), pr); err != nil {
+			return false, nil
+		}
+
+		return pr.Status.ResourceSchemaName != arsName, nil
+	})
+	if err != nil {
+		t.Fatalf("Failed to wait for PublishedResource status to be updated: %v", err)
 	}
 
-	newARSName := pr.Status.ResourceSchemaName
-	if newARSName == "" {
+	if pr.Status.ResourceSchemaName == "" {
 		t.Fatal("Expected PublishedResource status to contain ARS name, but value is empty.")
-	}
-
-	if newARSName == arsName {
-		t.Fatalf("Expected PublishedResource status to have been updated with new ARS name, but still contains %q.", arsName)
 	}
 }
 
@@ -214,7 +223,7 @@ func TestARSOnlyContainsSelectedCRDVersion(t *testing.T) {
 		theVersion    = "v1"
 	)
 
-	ctx := context.Background()
+	ctx := t.Context()
 	ctrlruntime.SetLogger(logr.Discard())
 
 	// setup a test environment in kcp
@@ -296,7 +305,7 @@ func TestMultiVersionCRD(t *testing.T) {
 	// force a non-standard order, because it should not matter for the sync
 	var selectedVersions = []string{"v2", "v1"}
 
-	ctx := context.Background()
+	ctx := t.Context()
 	ctrlruntime.SetLogger(logr.Discard())
 
 	// setup a test environment in kcp
@@ -386,7 +395,7 @@ func TestProjection(t *testing.T) {
 		originalVersion = "v1"
 	)
 
-	ctx := context.Background()
+	ctx := t.Context()
 	ctrlruntime.SetLogger(logr.Discard())
 
 	// setup a test environment in kcp
@@ -489,7 +498,7 @@ func TestNonCRDResource(t *testing.T) {
 		originalVersion = "v1"
 	)
 
-	ctx := context.Background()
+	ctx := t.Context()
 	ctrlruntime.SetLogger(logr.Discard())
 
 	// setup a test environment in kcp
