@@ -90,11 +90,9 @@ func TestSyncRelatedObjects(t *testing.T) {
 				Kind:       "Secret",
 				Object: syncagentv1alpha1.RelatedResourceObject{
 					RelatedResourceObjectSpec: syncagentv1alpha1.RelatedResourceObjectSpec{
-						Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-							Path: "metadata.name", // irrelevant
-							Regex: &syncagentv1alpha1.RegularExpression{
-								Replacement: "my-credentials",
-							},
+						Template: &syncagentv1alpha1.TemplateExpression{
+							// same fixed value on both sides
+							Template: "my-credentials",
 						},
 					},
 				},
@@ -143,11 +141,9 @@ func TestSyncRelatedObjects(t *testing.T) {
 				Kind:       "Secret",
 				Object: syncagentv1alpha1.RelatedResourceObject{
 					RelatedResourceObjectSpec: syncagentv1alpha1.RelatedResourceObjectSpec{
-						Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-							Path: "metadata.name", // irrelevant
-							Regex: &syncagentv1alpha1.RegularExpression{
-								Replacement: "my-credentials",
-							},
+						Template: &syncagentv1alpha1.TemplateExpression{
+							// same fixed value on both sides
+							Template: "my-credentials",
 						},
 					},
 				},
@@ -177,216 +173,168 @@ func TestSyncRelatedObjects(t *testing.T) {
 
 		//////////////////////////////////////////////////////////////////////////////////////////////
 
-		// {
-		// 	name:      "sync referenced Secret up into a new namespace",
-		// 	workspace: "sync-referenced-secret-up-namespace",
-		// 	mainResource: crds.Crontab{
-		// 		ObjectMeta: metav1.ObjectMeta{
-		// 			Name:      "my-crontab",
-		// 			Namespace: "default",
-		// 		},
-		// 		Spec: crds.CrontabSpec{
-		// 			CronSpec: "* * *",
-		// 			Image:    "ubuntu:latest",
-		// 		},
-		// 	},
-		// 	relatedConfig: syncagentv1alpha1.RelatedResourceSpec{
-		// 		Identifier: "credentials",
-		// 		Origin:     "service",
-		// 		Kind:       "Secret",
-		// 		Object: syncagentv1alpha1.RelatedResourceObject{
-		// 			RelatedResourceObjectSpec: syncagentv1alpha1.RelatedResourceObjectSpec{
-		// 				Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-		// 					Path: "metadata.name", // irrelevant
-		// 					Regex: &syncagentv1alpha1.RegularExpression{
-		// 						Replacement: "my-credentials",
-		// 					},
-		// 				},
-		// 			},
-		// 		},
-		// 		Destination: syncagentv1alpha1.RelatedResourceDestination{
-		// 			RelatedResourceDestinationSpec: syncagentv1alpha1.RelatedResourceDestinationSpec{
-		// 				Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-		// 					Path: "metadata.name", // irrelevant
-		// 					Regex: &syncagentv1alpha1.RegularExpression{
-		// 						Replacement: "my-credentials",
-		// 					},
-		// 				},
-		// 			},
-		// 			Namespace: &syncagentv1alpha1.RelatedResourceDestinationSpec{
-		// 				Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-		// 					Path: "metadata.name", // irrelevant
-		// 					Regex: &syncagentv1alpha1.RegularExpression{
-		// 						Replacement: "new-namespace",
-		// 					},
-		// 				},
-		// 			},
-		// 		},
-		// 	},
-		// 	sourceRelatedObject: corev1.Secret{
-		// 		ObjectMeta: metav1.ObjectMeta{
-		// 			Name:      "my-credentials",
-		// 			Namespace: "synced-default",
-		// 		},
-		// 		Data: map[string][]byte{
-		// 			"password": []byte("hunter2"),
-		// 		},
-		// 		Type: corev1.SecretTypeOpaque,
-		// 	},
+		{
+			name:      "sync referenced Secret up into a new namespace",
+			workspace: "sync-referenced-secret-up-namespace",
+			mainResource: crds.Crontab{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-crontab",
+					Namespace: "default",
+				},
+				Spec: crds.CrontabSpec{
+					CronSpec: "* * *",
+					Image:    "ubuntu:latest",
+				},
+			},
+			relatedConfig: syncagentv1alpha1.RelatedResourceSpec{
+				Identifier: "credentials",
+				Origin:     "service",
+				Kind:       "Secret",
+				Object: syncagentv1alpha1.RelatedResourceObject{
+					RelatedResourceObjectSpec: syncagentv1alpha1.RelatedResourceObjectSpec{
+						Template: &syncagentv1alpha1.TemplateExpression{
+							Template: "my-credentials",
+						},
+					},
+					Namespace: &syncagentv1alpha1.RelatedResourceObjectSpec{
+						Template: &syncagentv1alpha1.TemplateExpression{
+							Template: `{{ if eq .Side "kcp" }}new-namespace{{ else }}{{ .Object.metadata.namespace }}{{ end }}`,
+						},
+					},
+				},
+			},
+			sourceRelatedObject: corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-credentials",
+					Namespace: "synced-default",
+				},
+				Data: map[string][]byte{
+					"password": []byte("hunter2"),
+				},
+				Type: corev1.SecretTypeOpaque,
+			},
 
-		// 	expectedSyncedRelatedObject: corev1.Secret{
-		// 		ObjectMeta: metav1.ObjectMeta{
-		// 			Name:      "my-credentials",
-		// 			Namespace: "new-namespace",
-		// 		},
-		// 		Data: map[string][]byte{
-		// 			"password": []byte("hunter2"),
-		// 		},
-		// 		Type: corev1.SecretTypeOpaque,
-		// 	},
-		// },
+			expectedSyncedRelatedObject: corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-credentials",
+					Namespace: "new-namespace",
+				},
+				Data: map[string][]byte{
+					"password": []byte("hunter2"),
+				},
+				Type: corev1.SecretTypeOpaque,
+			},
+		},
 
-		// //////////////////////////////////////////////////////////////////////////////////////////////
+		//////////////////////////////////////////////////////////////////////////////////////////////
 
-		// {
-		// 	name:      "sync referenced Secret down into a new namespace",
-		// 	workspace: "sync-referenced-secret-down-namespace",
-		// 	mainResource: crds.Crontab{
-		// 		ObjectMeta: metav1.ObjectMeta{
-		// 			Name:      "my-crontab",
-		// 			Namespace: "default",
-		// 		},
-		// 		Spec: crds.CrontabSpec{
-		// 			CronSpec: "* * *",
-		// 			Image:    "ubuntu:latest",
-		// 		},
-		// 	},
-		// 	relatedConfig: syncagentv1alpha1.RelatedResourceSpec{
-		// 		Identifier: "credentials",
-		// 		Origin:     "kcp",
-		// 		Kind:       "Secret",
-		// 		Object: syncagentv1alpha1.RelatedResourceObject{
-		// 			RelatedResourceObjectSpec: syncagentv1alpha1.RelatedResourceObjectSpec{
-		// 				Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-		// 					Path: "metadata.name", // irrelevant
-		// 					Regex: &syncagentv1alpha1.RegularExpression{
-		// 						Replacement: "my-credentials",
-		// 					},
-		// 				},
-		// 			},
-		// 		},
-		// 		Destination: syncagentv1alpha1.RelatedResourceDestination{
-		// 			RelatedResourceDestinationSpec: syncagentv1alpha1.RelatedResourceDestinationSpec{
-		// 				Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-		// 					Path: "metadata.name", // irrelevant
-		// 					Regex: &syncagentv1alpha1.RegularExpression{
-		// 						Replacement: "my-credentials",
-		// 					},
-		// 				},
-		// 			},
-		// 			Namespace: &syncagentv1alpha1.RelatedResourceDestinationSpec{
-		// 				Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-		// 					Path: "metadata.name", // irrelevant
-		// 					Regex: &syncagentv1alpha1.RegularExpression{
-		// 						Replacement: "new-namespace",
-		// 					},
-		// 				},
-		// 			},
-		// 		},
-		// 	},
-		// 	sourceRelatedObject: corev1.Secret{
-		// 		ObjectMeta: metav1.ObjectMeta{
-		// 			Name:      "my-credentials",
-		// 			Namespace: "default",
-		// 		},
-		// 		Data: map[string][]byte{
-		// 			"password": []byte("hunter2"),
-		// 		},
-		// 		Type: corev1.SecretTypeOpaque,
-		// 	},
+		{
+			name:      "sync referenced Secret down into a new namespace",
+			workspace: "sync-referenced-secret-down-namespace",
+			mainResource: crds.Crontab{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-crontab",
+					Namespace: "default",
+				},
+				Spec: crds.CrontabSpec{
+					CronSpec: "* * *",
+					Image:    "ubuntu:latest",
+				},
+			},
+			relatedConfig: syncagentv1alpha1.RelatedResourceSpec{
+				Identifier: "credentials",
+				Origin:     "kcp",
+				Kind:       "Secret",
+				Object: syncagentv1alpha1.RelatedResourceObject{
+					RelatedResourceObjectSpec: syncagentv1alpha1.RelatedResourceObjectSpec{
+						Template: &syncagentv1alpha1.TemplateExpression{
+							Template: "my-credentials",
+						},
+					},
+					Namespace: &syncagentv1alpha1.RelatedResourceObjectSpec{
+						Template: &syncagentv1alpha1.TemplateExpression{
+							Template: `{{ if eq .Side "kcp" }}{{ .Object.metadata.namespace }}{{ else }}new-namespace{{ end }}`,
+						},
+					},
+				},
+			},
+			sourceRelatedObject: corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-credentials",
+					Namespace: "default",
+				},
+				Data: map[string][]byte{
+					"password": []byte("hunter2"),
+				},
+				Type: corev1.SecretTypeOpaque,
+			},
 
-		// 	expectedSyncedRelatedObject: corev1.Secret{
-		// 		ObjectMeta: metav1.ObjectMeta{
-		// 			Name:      "my-credentials",
-		// 			Namespace: "new-namespace",
-		// 		},
-		// 		Data: map[string][]byte{
-		// 			"password": []byte("hunter2"),
-		// 		},
-		// 		Type: corev1.SecretTypeOpaque,
-		// 	},
-		// },
+			expectedSyncedRelatedObject: corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-credentials",
+					Namespace: "new-namespace",
+				},
+				Data: map[string][]byte{
+					"password": []byte("hunter2"),
+				},
+				Type: corev1.SecretTypeOpaque,
+			},
+		},
 
-		// //////////////////////////////////////////////////////////////////////////////////////////////
+		//////////////////////////////////////////////////////////////////////////////////////////////
 
-		// {
-		// 	name:      "sync referenced Secret up from a foreign namespace",
-		// 	workspace: "sync-referenced-secret-up-foreign-namespace",
-		// 	mainResource: crds.Crontab{
-		// 		ObjectMeta: metav1.ObjectMeta{
-		// 			Name:      "my-crontab",
-		// 			Namespace: "default",
-		// 		},
-		// 		Spec: crds.CrontabSpec{
-		// 			CronSpec: "* * *",
-		// 			Image:    "ubuntu:latest",
-		// 		},
-		// 	},
-		// 	relatedConfig: syncagentv1alpha1.RelatedResourceSpec{
-		// 		Identifier: "credentials",
-		// 		Origin:     "service",
-		// 		Kind:       "Secret",
-		// 		Object: syncagentv1alpha1.RelatedResourceObject{
-		// 			RelatedResourceObjectSpec: syncagentv1alpha1.RelatedResourceObjectSpec{
-		// 				Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-		// 					Path: "metadata.name", // irrelevant
-		// 					Regex: &syncagentv1alpha1.RegularExpression{
-		// 						Replacement: "my-credentials",
-		// 					},
-		// 				},
-		// 			},
-		// 			Namespace: &syncagentv1alpha1.RelatedResourceObjectSpec{
-		// 				Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-		// 					Path: "metadata.name", // irrelevant
-		// 					Regex: &syncagentv1alpha1.RegularExpression{
-		// 						Replacement: "other-namespace",
-		// 					},
-		// 				},
-		// 			},
-		// 		},
-		// 		Destination: syncagentv1alpha1.RelatedResourceDestination{
-		// 			RelatedResourceDestinationSpec: syncagentv1alpha1.RelatedResourceDestinationSpec{
-		// 				Reference: &syncagentv1alpha1.RelatedResourceObjectReference{
-		// 					Path: "metadata.name", // irrelevant
-		// 					Regex: &syncagentv1alpha1.RegularExpression{
-		// 						Replacement: "my-credentials",
-		// 					},
-		// 				},
-		// 			},
-		// 		},
-		// 	},
-		// 	sourceRelatedObject: corev1.Secret{
-		// 		ObjectMeta: metav1.ObjectMeta{
-		// 			Name:      "my-credentials",
-		// 			Namespace: "other-namespace",
-		// 		},
-		// 		Data: map[string][]byte{
-		// 			"password": []byte("hunter2"),
-		// 		},
-		// 		Type: corev1.SecretTypeOpaque,
-		// 	},
+		{
+			name:      "sync referenced Secret up from a foreign namespace",
+			workspace: "sync-referenced-secret-up-foreign-namespace",
+			mainResource: crds.Crontab{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-crontab",
+					Namespace: "default",
+				},
+				Spec: crds.CrontabSpec{
+					CronSpec: "* * *",
+					Image:    "ubuntu:latest",
+				},
+			},
+			relatedConfig: syncagentv1alpha1.RelatedResourceSpec{
+				Identifier: "credentials",
+				Origin:     "service",
+				Kind:       "Secret",
+				Object: syncagentv1alpha1.RelatedResourceObject{
+					RelatedResourceObjectSpec: syncagentv1alpha1.RelatedResourceObjectSpec{
+						Template: &syncagentv1alpha1.TemplateExpression{
+							Template: "my-credentials",
+						},
+					},
+					Namespace: &syncagentv1alpha1.RelatedResourceObjectSpec{
+						Template: &syncagentv1alpha1.TemplateExpression{
+							Template: `{{ if eq .Side "kcp" }}{{ .Object.metadata.namespace }}{{ else }}other-namespace{{ end }}`,
+						},
+					},
+				},
+			},
+			sourceRelatedObject: corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-credentials",
+					Namespace: "other-namespace",
+				},
+				Data: map[string][]byte{
+					"password": []byte("hunter2"),
+				},
+				Type: corev1.SecretTypeOpaque,
+			},
 
-		// 	expectedSyncedRelatedObject: corev1.Secret{
-		// 		ObjectMeta: metav1.ObjectMeta{
-		// 			Name:      "my-credentials",
-		// 			Namespace: "default",
-		// 		},
-		// 		Data: map[string][]byte{
-		// 			"password": []byte("hunter2"),
-		// 		},
-		// 		Type: corev1.SecretTypeOpaque,
-		// 	},
-		// },
+			expectedSyncedRelatedObject: corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-credentials",
+					Namespace: "default",
+				},
+				Data: map[string][]byte{
+					"password": []byte("hunter2"),
+				},
+				Type: corev1.SecretTypeOpaque,
+			},
+		},
 
 		//////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -416,9 +364,9 @@ func TestSyncRelatedObjects(t *testing.T) {
 								},
 							},
 							Rewrite: syncagentv1alpha1.RelatedResourceSelectorRewrite{
-								// TODO: Use template instead of regex once that is implemented.
-								Regex: &syncagentv1alpha1.RegularExpression{
-									Replacement: "my-credentials",
+								Template: &syncagentv1alpha1.TemplateExpression{
+									// same fixed name on both sides
+									Template: "my-credentials",
 								},
 							},
 						},
@@ -453,11 +401,80 @@ func TestSyncRelatedObjects(t *testing.T) {
 				Type: corev1.SecretTypeOpaque,
 			},
 		},
+
+		//////////////////////////////////////////////////////////////////////////////////////////////
+
+		{
+			name:      "find Secret based on templated label selector",
+			workspace: "sync-templated-selected-secret-up",
+			mainResource: crds.Crontab{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-crontab",
+					Namespace: "default",
+				},
+				Spec: crds.CrontabSpec{
+					CronSpec: "* * *",
+					Image:    "ubuntu:latest",
+				},
+			},
+			relatedConfig: syncagentv1alpha1.RelatedResourceSpec{
+				Identifier: "credentials",
+				Origin:     "service",
+				Kind:       "Secret",
+				Object: syncagentv1alpha1.RelatedResourceObject{
+					RelatedResourceObjectSpec: syncagentv1alpha1.RelatedResourceObjectSpec{
+						Selector: &syncagentv1alpha1.RelatedResourceObjectSelector{
+							LabelSelector: metav1.LabelSelector{
+								MatchLabels: map[string]string{
+									// include some nasty whitespace
+									`  {{ list "fi" "nd" | join "-" }} `: `
+{{ lower "ME" }}
+ `,
+								},
+							},
+							Rewrite: syncagentv1alpha1.RelatedResourceSelectorRewrite{
+								Template: &syncagentv1alpha1.TemplateExpression{
+									// same fixed name on both sides
+									Template: "my-credentials",
+								},
+							},
+						},
+					},
+				},
+			},
+			sourceRelatedObject: corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "unknown-name",
+					Namespace: "synced-default",
+					Labels: map[string]string{
+						"fi-nd": "me",
+					},
+				},
+				Data: map[string][]byte{
+					"password": []byte("hunter2"),
+				},
+				Type: corev1.SecretTypeOpaque,
+			},
+
+			expectedSyncedRelatedObject: corev1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "my-credentials",
+					Namespace: "default",
+					Labels: map[string]string{
+						"fi-nd": "me",
+					},
+				},
+				Data: map[string][]byte{
+					"password": []byte("hunter2"),
+				},
+				Type: corev1.SecretTypeOpaque,
+			},
+		},
 	}
 
 	for _, testcase := range testcases {
 		t.Run(testcase.name, func(t *testing.T) {
-			ctx := context.Background()
+			ctx := t.Context()
 
 			// setup a test environment in kcp
 			orgKubconfig := utils.CreateOrganization(t, ctx, testcase.workspace, apiExportName)
@@ -481,8 +498,8 @@ func TestSyncRelatedObjects(t *testing.T) {
 					},
 					// These rules make finding the local object easier, but should not be used in production.
 					Naming: &syncagentv1alpha1.ResourceNaming{
-						Name:      "$remoteName",
-						Namespace: "synced-$remoteNamespace",
+						Name:      "{{ .Object.metadata.name }}",
+						Namespace: "synced-{{ .Object.metadata.namespace }}",
 					},
 					Projection: &syncagentv1alpha1.ResourceProjection{
 						Group: "kcp.example.com",
